Fixed: The initial linux commit hangs the ui.
Now we just trim it to 1024*1024 lines.
Number reached is totally arbitrary, but appending really huge number of
lines to QTextEdit seems to hang it. A more approriate solution is needed.
Perhaps i need to replace the diff textbox with a proper diff model/view
based arch. Since we need to do that for commit tools anyway.
Until then atleast this will prevent a crash.